requirements.txt
functions-framework==3.*
slackclient==2.9.4
flask==2.1.0
requests==2.28.1